requirements.txt
audioread==3.0.1
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
chatterbox-tts==0.1.3
conformer==0.3.2
Cython==3.1.5
decorator==5.2.1
Deprecated==1.2.18
diffusers==0.29.0
distlib==0.4.0
einops==0.8.1
filelock==3.19.1
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.3
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
jaconv==0.4.0
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
MarkupSafe==3.0.3
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.2
networkx==3.2.1
nodeenv==1.9.1
numba==0.60.0
numpy==2.0.2
onnx==1.19.1
packaging==25.0
pillow==11.3.0
pkuseg==0.0.25
platformdirs==4.4.0
pooch==1.8.2
pre_commit==4.3.0
protobuf==6.33.0
pycparser==2.23
pykakasi==2.3.0
PyYAML==6.0.3
regex==2025.10.22
requests==2.32.5
resemble-perth==1.0.1
s3tokenizer==0.2.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
soundfile==0.13.1
soxr==1.0.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.20.3
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.15.0
urllib3==2.5.0
virtualenv==20.35.3
wrapt==1.17.3
zipp==3.23.0
